// src/interfaces/User.ts
export default interface User {
  fotoDePerfil: string;
  apellidoPaterno: string;
  apellidoMaterno: string;
  primerNombre: string;
  segundoNombre: string;
  nombreUsuario: string;
  password: string;
  fechaNacimiento: string;
  lugarNacimiento: string;
  sexo: string;
  carnetIdentidad: string;
  direccionZona: string;
  telefono: string;
  celular: string;
  email: string;
  alergiaMedicamento: string;
  estado: boolean;
}